Add company.ts
