+++ src/sections/Landing.tsx
@@ -10,14 +10,14 @@ import LandingDonutChart from '../components/Charts/LandingDonutChart'
 const Landing: FC = () => {
   const textData = data.listItems
   const landingTextStyle = clsx(
-    'font-semibold underline decoration-[#FFA902] decoration-2 underline-offset-4',
+    'font-semibold underline decoration-clementine decoration-2 underline-offset-4',
   )
 
   return (
     <>
       <CapitalizedTitle />
-      <div className="mx-14 flex justify-center py-10">
-        <div className="donut place-content-center px-14 pt-6">
+      <div className="mx-14 flex flex-row justify-center py-10">
+        <div className="px-14 pt-6">
           <LandingDonutChart />
         </div>
         <div>
@@ -31,7 +31,7 @@ const Landing: FC = () => {
             <span>{textData[4]}</span>
             <span className={landingTextStyle}>{textData[5]}</span>
           </p>
-          <div className="flex gap-12 pt-4">
+          <div className="mt-4 flex flex-row gap-12">
             <PrimaryButton text="More About Fossil Fuels" link="/fossil-fuel" />
             <PrimaryButton text="Learn About Investment" link="/investments" />
           </div>
