+++ src/server/api/root.ts
@@ -1,5 +1,6 @@
 import { createTRPCRouter } from './trpc'
-import { companyRouter, exampleRouter } from './routers/example'
+import { exampleRouter } from './routers/example'
+import { companyRouter } from './routers/company'
 import { esgRouter } from './routers/esg'
 
 /**
