+++ src/server/api/routers/example.ts
@@ -15,14 +15,3 @@ export const exampleRouter = createTRPCRouter({
     return 'Secret message: SECS is the best team!'
   }),
 })
-
-export const companyRouter = createTRPCRouter({
-  countByInvestment: publicProcedure.query(({ ctx }) => {
-    return ctx.prisma.company.groupBy({
-      by: ['sector'],
-      _count: {
-        sector: true,
-      },
-    })
-  }),
-})
